1.6

+++ functions.py
@@ -1,7 +1,7 @@
 def promedio_estudiante(calificaciones):
     if len(notas) == 0:
         return 0.0
-    return float(sum(calificaciones) / len(calificaciones))
+    return sum(calificaciones) / len(calificaciones)
 notas= []
 promedio= promedio_estudiante(notas)
 print(promedio)
